Remove deprecated reserveMapFileReader
Remove deprecated reserveMapFileReader from IteratorEnvironment. This
change affects the behavior of the MapFileIterator's deepCopy, using the
implementation from the DefaultIteratorEnvironment that opens a new file
reader on the given map file instead of the TabletIteratorEnvironment's
implementation that would retrieve a reference to an already open file
reader that was cached.

For context, map files have fallen out of favor, and RFiles have been in
use for a long time as their replacement (since at least Accumulo 1.3).
I do not know if this MapFileIterator code is likely to even work, as I
don't know of anybody still using map files. The support for reading map
files should probably be removed in their entirety, the property to
control the file extension removed (or at least, currently restricted to
only accept RFile.EXTENSION as valid), and upgrade should be blocked
until all map files are compacted away. This commit is much more
narrowly focused than that, but should help move towards that end goal.