Add CI check to detect Thrift changes
Add a GitHub Actions check that re-generates the Thrift generated code,
and compares the alterations to what's committed to the repo. The check
will fail if Thrift generates something that isn't checked in. This
ensures that we can ignore generated code changes when reviewing pull
requests, because if the generated code contained anything unexpected,
this check would make it obvious.

This fixes #3095