Fix issue #3371 - tranform Error to RunTimeException in scans
Scan issues that throw Errors were not handled and caused the process to
terminate.  The errors can be caused by user code, and should kill
the scan thread, not the entire process.

Translating Errors to RuntimeExceptions allows the thread
exception handler to process the exception, but prevents killing the
process.

Adds a test iterator that can throw Errors and validates an exception
on scan propigate to clinet, but process is not killed.