Disable MapFile in 1.10 and 2.1 patch releases
Drop MapFile support (#3359) by partially backporting PR #3378, and
making minimally disruptive changes to avoid use of MapFiles for 1.10.4
and 2.1.1.

As noted in #3378, MapFile support is already broken and has been for a
long time. This change will cause an explicit and detectable failure,
rather than a silent one, if a MapFile is attempted to be used.