fixes race condition in metadata consistency check
While looking into #3386 I noticed the Accumulo metadata consistency
check was incrementing a counter in the incorrect place.  It should
increment before writing to the metadata table, but it does not.
This could cause the check to report false postives.  The false
positive in the case would be transient and should not repeat on
subsequent checks.

Also noticed a redundant check when deciding if the file should be
added to the set of in memory files.  AFICT this redundant check
is harmless, but it could cause problems for future changes.