Start using TabletFile instead of String for referencing data files
This commit moves away from using a String or Path for referring to data
files when reading/writing and instead uses a TabletFile so that range
information can be more easily added in the future. New classes have
been introduced called AbstractTabletFile and UnassignedTabletFile.
UnassignedTabletFile represents a TabletFile that is not yet part of a
Table/Tablet yet but still needs to be written or read from such as in
the Bulk Import case, WAL, testing, etc.