Improve some exception handling
Use a more specific RuntimeException type, for a large number of
appropriate situations. Clean up some unnecessary wrapping of
RuntimeExceptions with a new RuntimeException.

Also, try to catch more specific checked exceptions, when appropriate,
instead of catching (Exception) and do the same in non-public API
methods where Exception was being thrown, forcing catch blocks to handle
a wider number of exceptions than were actually thrown.

This change represents an incremental improvement, and does not try to
be an exhaustive fix of all these similar issues. There may be other
opportunities for similar improvements that are not included in this
change and can be done in a subsequent change. Rather, this more
narrowly fixes the occurrences originally included as part of abandoned
PR #2762 to address spotbugs warnings that no longer occur on the latest
version of spotbugs.