Set finished time for GC cycles even if Exception
Always set the finished time for GC cycles, even if an exception
occurred in the cycle. This prevents a GCStatus object containing a
started time, but the finished time staying the default value of 0.

Ideally, this wouldn't be a problem if we used the setter/getter methods
on Thrift objects so that we can tell the difference between a 0 that
has been set and one that is 0 just because it's a primitive type that
defaults to 0 when unset. Thrift tracks which fields are set using
isSetFieldname methods, but since we're not consistently using those,
and just grabbing the value of the field without checking if it has been
set, we need to make sure that we've set it to something sensible.

I believe this fixes #3374