Moved automatic splits to the manager (#3382)
The Manager now scans tablets looking for ones that need to split.  When
it finds one it queues it on a thread pool for inspection of files to
find a split point.  When split points are found a FATE operation is
initiated to do the splits.  The FATE operation uses conditional
mutations to synchronize access to the tablet metadata.  The FATE
operation creates one or more new tablets in the metadata table. Unlike
the the old split code which could only split a tablet into two, this
new code can split a tablet into however many are needed in a single
operation.  This makes responding to a change in the split threshold
much faster.

The old code used to shorten split points, this change carries that
behavior forward in the new code to find multiple split points.

This change does not include moving user intiated splits into the
manager, that will be done in a follow on change.