Modified TabletGroupWatcher to use TabletMetadata (#3409)
Removed TabletStateChangeIterator and TabletLocationState objects
and replaced them a TabletManagementIterator that returns TabletManagement
objects. TabletManagement objects contain the TabletMetadata (same as
returned by Ample) and a set of actions that are needed for the Tablet
(location update, split, merge, compaction, etc.). There were
some places where TabletLocationState was being used as a way to pass
tablet metadata to a method (UnloadTabletHandler and some tests).
For these cases, I created TabletMetadataImposter which extends
TabletMetadata and sets its members using the variables passed to the
constructor vs the normal way of creating TabletMetadata (parsing
metadata table K/V pairs). The new TabletManagementIterator and
TabletManagement objects will allow the TabletGroupWatcher threads in
the Manager take the necessary action on tablets, but distribute the
computational load for determining the actions to the TabletServers.