Introduce new GC trash property, move duplicate methods to new class
This commit deprecates GC_TRASH_IGNORE in favor of GC_USE_TRASH. The
description for the deprecated property and the way that it was used
is confusing as it's written and used in the negative. If you want to
use the Trash, you have to set it the property to "false". The new
property works in the affirmative. Set it to "true" if you want to use
the Trash, "false" if you don't. A third value, "bulk_imports_only" means
that you only want to send bulk import files to the Trash.