reduces metadata reads and caches hosting request in client tablet cache
The code to host a tablet was reading from the metdata table to gather a
tablets hosting goal. The hosting goal was already present in the cache.
Used the value from the cache instead of reading from the metadata
table again.

Also added the hosting request column to the tablet cache and removed a
specialized caffine cache that was related to this.  Moving it into the
tablet cache will help avoid request when other client processes have
already made the request, the specialized caffine cache was not enabling
this case.

fixes #3304