moves hosting request processing from tserver to manager (#3437)
Processing of hosting request was moved from the tserver to manager so
that the tablet group watcher could be notified.  This considerably
reduces the delay in getting an ondemand tablet hosted.

Introduced a cache to avoid reprocessing hosting request for the same
tablet in a short time period.

Modified the update of the hosting request column to use a conditional
mutation.  This avoids a metadata read before write (the read is now
done by the mutation conditions) and avoids race conditions.

Moving the hosting request to the manager could introduce scaling
problems, however the cache can help with this. The reduced latency for
hosting ondemand tablet is needed, so if there are scaling problems a
solution will need to be found that has low latency for hosting.