Remove the deprecated -s (safemode) option from ConfigOpts (#3444)
The ConfigOpts used -s option for accumulo-gc's safemode.
To allow utilities to use -s, remove the deprecated option and
the legacy options parsing that's not needed.

This was only ever preserved to issue a slightly more verbose
error message when using a legacy flag that had been removed.
The option itself was only ever used by the accumulo-gc and
placed in here as a hidden option when consolidating GCOpts
with its parent class.

Removing this does not change any behavior for users, but
results in a slightly less informational message if the user
does try to use this invalid option. But, that's an acceptable
trade-off so this option is not held reserved, blocking its use
by other command-line tools that might need it.