Refactor StoredTabletFile to contain a TabletFile (#3417)
This commit refactors StoredTabletFile so it extends AbstractTabletFile
and becomes a sibling of TabletFile. StoredTabletFile now contains a
TabletFile instead of extending TabletFile.

These changes solve the problem of not being able to create a
hashcode/equals implementation for StoredTabletFile has many places in
the code treated these objects as equal. Establishing a composite
relationship creates a clear distinction and allows proper equality and
comparison checks.

This fixes #2830