Add an optional range to Tablet files and fenced reader (#3418)
This commit adds a Range to TabletFile that can be used to fence
off the data read from an RFile. A new iterator has been created
that will only return rows in the file that match the given Range.
If a Range is set that has a null start/end key then the file
will not be fenced. 

The changes here allow treating the same RFile as a different
TabletFile if the Range is different which will make no chop merges
possible in the future. 

Co-authored-by: Keith Turner <kturner@apache.org>