Partially moves compactions into the manager
This commit moves system compactions into the manager. User compactions
are not done yet. The TabletManagmentIterator now finds tablets that
need compaction by calling the compaction dispatcher and compaction
planner plugins.  Any tablets that needs compaction is returned to the
manager where it runs the plugins again to generate compaction jobs. The
compaction jobs are placed in a new bounded priority queue. The queue is
bounded in size to prevent exhausting manager memory, it keeps the Top N
jobs with the highest priority.  The CompactionCoordinator pulls jobs
out of then new queue when compactors request work.  A conditional
mutation is used in the compaction coordinator to atomically reserve the
files for compaction. When the compactor reports the job is done the
coordinator atomically updates the tablet metadata using a conditional
mutation and if the tablet is hosted its asked to refresh its
metadata.

User compactions in the manager still need to be implemented.  Also many
loose ends still need to be implemented for system compactions, these
changes will be done in follow on commits.