fixes bug where bulk import could cause compcation to hang
This commit fixes a bug where a bulk import into a tablet that could not
split would leave the tablet in bad state where it could not compact.
The cause of the bug was the split code calling Tablet.initiateClose()
which closed the tablets compactable object.  Then split code would
decided it did not want to close the tablet and reset Tablet.closeState
to OPEN, however the compactable object would be left closed.

This bug existed before #3249, however the changes in #3249 exposed a
reliable path to the bug via bulk import.  Before #3249 I suspect the
only path to access the bug was a race conditions where an unsplittable
file was added at just the right time while the split code was running.

This particular bug with the closing of the compacatable and subsequent
reopening of the tablet does not exists in 1.X.  However the sketchy
code that initiates close and then reopens a tablet  does exists in 1.X.
If that code has other problems, then #3249 related changes in 1.10 may
provide a new path to accessing those problems.

This commit also adds a test that was able to reproduce the problem.

fixes #3491