Add TSERV_GROUP_NAME, ability to balance within tserver group
Added TSERV_GROUP_NAME property and refactored how groups
are determined within the server processes. Modified LiveTServerSet
to pull the group from the ServiceLockData and add it to the
TServerInstance. Modified the TabletGroupWatcher to use the
group information in the TServerInstance to determine if the
tablet is currently hosted on a TServer that is not in the
correct group