StoredTabletFile now lazy loads ReferencedTabletFile (#3502)
This improves performance because creating a ReferencedTabletFile does a
lot of validation which can be slow due to the calls to Path.getParent()
if there are several files created. Filename validation was also removed
from the constructor as well to further improve performance.

Removing/delaying the validation checks should be safe because a
StoredTabletFile references an existing Tablet file that was already
stored in Metadata so the Path has already been validated before. Also,
the ReferencedTabletFile is not needed often so it makes sense to delay
the creation/validation until it's actually needed.