Verify hostingGoals in table with mixed goals
This PR adds a suggested additional test for the setting of
TabletHostingGoals in TableOperationsImpl.java.

Specifically, the new tests verifies that when a table that already
contains mixed TabletHostingGoals splits further, the new splits retain
the hostingGoal of the tablet from which they were split. The new test,
testGetHostingGoals_StaggeredSplits, does the following:

- create table
- add split 'm'
- set the hostingGoal of the last tablet to ALWAYS, leaving the first
  tablet with the default ONDEMAND
- add additional splits 'd' and 's'
- verify the first two tablets now have a hostingGoal of ONDEMAND
- verify the last two tablets, which were split from an existing
  tablet with ALWAYS as the goal, both have a goal of ALWAYS