Modified MiniAccumuloCluster configuration to include cluster config (#3533)
Created an object used by MiniAccumuloConfigImpl that contains the
number of Compactions, ScanServers, and TabletServers for the default
resource group. This object allows test writers to change the defaults
and add other resource groups for their test. When MAC starts, it starts
all of the Compactors, ScanServers, and TabletServers for all of the groups.
Resource groups can be added to the configuration in the middle of a test
and then calling start(ServerType) will start up the new processes.