Backported MANAGER_TABLET_GROUP_WATCHER_INTERVAL property (#3542)
TabletGroupWatcher prior to this change was hardcoded to
wait for 60s between loop iterations looking for tablets that needed
some type of maintenance. Created a property allows the user to
configure that potentially making the Manager more responsive
to tablets being unassigned or unbalanced. Also created a property
whose value is used as the TTL value for recovery write-ahead logs
in a Cache.

Fixes #3541