Reliably refresh tablets after updating tablet files for a compaction. (#3563)
This commit modifies the compaction coordinator to reliably refresh
hosted tablets in memory metadata after a compaction.  The RPC request
to a hosted tablet asking it to refresh happens after updating a tablets
metadata for a compaction. The changes in the PR ensure that if the
manager process dies after making the metadata update but before it
makes the RPC that this will happen when the manager restarts.

The implementation for this relies on a new ~refresh sections in the
metadata table.  System compactions on hosted tablets will write a
~refresh entry before trying to update a tablets metadata.  This entry
allows the compaction commit process to be fault tolerant.

Javadoc was added to CompactionCoordinator.compactionCompleted()
describing the overall process of committing a compaction.

This commit also refactored the refresh RPC code in bulk import so that
it could also be used by compactions.

This commit also made the tablet metadata update for a compaction add
scan entries for removed files when a tablet is hosted.  This was done
to prevent the Accumulo GC from removing compacted files that may be in
use by an immediate scan.

fixes #3462