Makes user compaction metadata independent.
Before this change user initiated compactions stored config in a single
place in zookeeper and had a single counter in each tablet.  This system
made it impossible for concurrent user initiated compactions to run on
the same table with different configuration.  This commit adds
compaction config storage and tablet metadata that has 1:1
correspondence with the fate operation driving a user initiated
compaction.  This allows each user initiated compaction to be
independent.  This new system also avoid some odd race conditions with
tablet metadata updates that existed with old per tablet compaction
counter.  Instead of a counter each tablet now conceptually has a set of
the fate transaction ids that have completed a compaction.

See the new documention in TableOperations.compact() for an example.

fixes #3517