Add TSERV_GROUP_NAME, ability to balance within tserver group (#3496)
Added TSERV_GROUP_NAME property and refactored how groups
are determined within the server processes. Modified LiveTServerSet
to pull the group from the ServiceLockData and add it to the
TServerInstance. Modified the TabletGroupWatcher to use the
group information in the TServerInstance to determine if the
tablet is currently hosted on a TServer that is not in the
correct group. Modified TableLoadBalancer to add a custom
property that allows the user to configure which tablet servers
a table should be assigned/balanced within.

Related to #3459