WIP - no chop merge proof of concept
This commit disables chop compactions on merge and modifies the tablet
metadata update in TabletGroupWatcher to fence of each tablet file to
its tablet range as its copied to the highest tablet on merge. This
prevents data from coming back on merge as scans will be fenced to the
provided range on the metadata updates.