Removes race conditions between tablet refresh and minor compaction
This commit moved the tablet twoards using a TabletMetadata object as
the source of the tablets files instead of tracking the files itself.
The move twoards using TabletMetadata support the changes to fix the
race conditions with tablet refresh.  Committing a minor compaction to
walogs and the metadata table is now mutually exclusive with refreshing
a tablets metadata.  Scans should always see a consistent view of the in
memory maps a tablet has and the files it has after this change.  Before
this change a tablet scan could see the file produced by a minor
compaction and the in memory map that was the source of that file.
Scans should only see one or the other.

fixes #3625