Create getTabletInformation API method
This PR creates a new public API method, getTabletInformation, which
returns a stream of TabletInformation objects containing information for
each tablet within a provided table. The method can be accessed from the
code or JShell via the TableOperations object.

A new TabletInformation class is used to hold information for an
individual tablet.

The 'listtablets' command has been modified to use this new method
rather than using non-public API. The ListTabletsCommandTest was updated
to reflect the new method. A new IT test has been added to ShellServerIT
which tests 'listtablets' and effectively tests the new API since the
method is used to obtain the information for 'listtablets'.