lowers number of files in compaction IT test
The test CompactionIT.testErrorDuringUserCompaction() was timing out. The
test would create a tablet with 1000 files.  These would compact in
batches of 10 files at a time with a 4 or 5 seconds between each
compaction.  So this could take 400 to 500 seconds.  The few second
delay between compactions comes from waiting for the manager to scan the
metadata table again. When compactions ran in the tablet server it would
immediately start another compaction when one finished.

This commit changes the test to create 100 files in the tablet instead
of 1000.