Improvements and cleanup to ExternalCompaction ITs and related code (#3661)
Modified ExternalCompactionITs to remove the startup/shutdown of Compactors
that are defined in the compaction services created by the MAC configuration
callback. MAC now starts these Compactors based on the compaction service
configuration. Made some modifications to ExternalCompactionITs to reduce the
load on ZK as some calls where being made to find the compaction coordinator
address inside of a loop, and that wasn't necessary. Made a modification to the
CompactionCoordinator so that it won't try to update tablet metadata for a table
that has been deleted. Made a modification to the Manager so that it only advertises
the FATE and CompactionCoordinator Thrift service after the Manager has started up.