Attempt to prevent half-closed Tablet due to failing minc
Prior to this change Tablet.initiateClose would close the
compactable, wait for the current minor compaction to finish,
and then kick off another minor compaction. In the case where
minor compactions are failing, the compactable will get closed
and the call to wait for the current minor compaction to finish
will hang indefinitely leaving the Tablet in a half-closed
state. Tablet.initiateClose is called when either the
TabletServer is closing the Tablet (due to migration or
shutdown) or on a Tablet split. Therefore, a failing minor
compaction will prevent Tablet migration and split or normal
TabletServer shutdown.

This change adds some logic at the start of Tablet.initiateClose
to try and detect a currently or previously failing minor
compaction. In these cases an exception is thrown before the
compactable is closed to prevent the Tablet from being in a
half-closed state. This prevents the Tablet from being closed
until the cause for the failing minor compaction is corrected.
Causes could be a bad iterator applied at minor compaction or a
bad table configuration option (like classloader context). When
the cause of the failing minor compaction is corrected, then the
Tablet should be able to be closed normally allowing normal
migration, split, and TabletServer shutdown.

Fixes #3674

Co-authored-by: Ed Coleman <edcoleman@apache.org>
Co-authored-by: dtspence <dtspence@users.noreply.github.com>