Validate that classloader context is valid when property is set (#3678)
Added method to ContextClassLoaderFactory that is called when
the table.context.class.loader property is set to validate that the
context name is valid. Default implementation returns true if the
implementation returns a non-null ClassLoader and does not throw
an exception. The new method is also called from
ClassLoaderUtil.getClassLoader(String) and will return the non-context
aware Accumulo ClassLoader if isValid returns false.


Co-authored-by: Ed Coleman <edcoleman@apache.org>