Avoids full table scans in TabletGroupWatcher in some situations.
The manager has an internal class called EventCoordinator that is used
to signal that something happened in the manager.  When anything
signaled an event happened this would cause the tablet group watcher to
scan the entire metadata table.  This commit makes a few changes to
improve this.

First it added the ability to EventCoordinator to signal with more
specific information, like a something happened with a table or a
tablet.

Second the TabletGroupWatcher was adapted to react to this more specific
signaling with narrower scans of the metadata table.  If a signal is
made that something changed with a table, then only that tables metadata
will be scanned.  The TabletGroupWatcher will still do full scans
periodically (based on configuration) or when a something signals a full
scan is needed.

A new IT was added that creates a 1 million splits in a table and then
reads and writes to around 100 random tablets.  The tablets are ondemand
so must be brought online for the read and write.  If doing bringing
them onling takes too long (because a full tables scan is done) then the
test might timeout and fail.