Fix FencedReader getLastKey() and splits
This update fixes getLastKey() to correctly return the last included key
of the Range for a TabletFile, and not the end key of the range which
will always fall outside the valid range.

Because the endKey is always required to be exclusive, the actual end
key of the range can't be returned from getLastKey() as that key is not
part of the range. Certain constructors for the Range() object already
append a 0x00 byte to the end of the key when converting from inclusive
to exclusive so we can require that this exists and then we can strip
that byte and get the last included key in the range to return.