Adjust shell to indicate how re-authtication occurs
This PR adds presents additional information to the shell when the
re-authentication is automatically performed using the value from
accumulo-client.properties.

Once merged, the update will be applied to 3.x where it corrects an
situation where an endless loop can be triggered when attempting
re-authentication if the accumulo-client.properties file password
is altered after initial setup.

See #3485 for more detail.