No Chop Merges (#3640)
This commit disables chop compactions on merge and modifies the tablet
metadata update in TabletGroupWatcher to fence of each tablet file to
its tablet range as its copied to the highest tablet on merge. 
By fencing off the files in a tablet on merge chop compactions are no 
longer necessary as this prevents data from coming back on merge 
as scans will be fenced to the provided range on the metadata updates.