Remove usages of the Writer class
Remove usages of the Writer class, used for single-mutation writes.
Replace its usages with the BatchWriter.

This fixes #3735 for 2.1.3 for all usages except the deprecated
Replication code, which is removed in 3.x

This code was originally authored by @keith-turner as part of PR #3733,
but separated out to simplify viewing the remaining changes in that PR,
which directly addresses the bug in #3721

Co-authored-by: Keith Turner <kturner@apache.org>