fixes bug w/ deletion of compaction thread pool
Fixes a bug where a change in compaction config that deleted a
compaction thread pool would leave any tablets compacting on that thread
pool in a bad state. The problem was caused by shutdownNow being called
on the thread pool which could cause an interrupted exception during
commiting the compaction.

This change allows compactions running on the thread pool to complete
while canceling any compactions queued on the thread pool.

fixes #3749