avoids creating server side threads when canceling failed write
This commit is a follow up to #3733. When a failure happens on a batch
write, the batch writer will continually try to close the write session.
Each time it tried to close the session it created a thread that may
wait on the tserver trying to reserve the session which takes up a server
side thrift thread pool slot.  After the 2 min default connection
timeout it will try to close the session again, creating another thread
on the tserver that is trying to close the session (the previous threads
are still there).  This will only happen when something is preventing
the the session from closing.  So in the case where many clients are
doing this it has the potential to make an existing problem worse.

This change introduces a new tablet server RPC named cancelUpdate. The
new method does not try to reserve the session.  If the session is
reserved it returns false, if not it deletes it and returns true.  The
client will retry with backoff calling this new RPC until it returns
true.  This new RPC will not leave threads waiting on the tserver. Since
this change was made in a bug fix release, it can detect if the new RPC
method does not exists and fall back to the existing closeUpdate RPC.