reverts changes from #3747 that caused tablet split to fail
Reverts a validation to prevent deadlock that was being covered by the
code path to close a tablet.  #3759 was opened to investigate this.
This commit removes the new check as the code path that could cause
deadlock has existed for a long time and may never actually cause
deadlock.

fixes #3757