No chop merge and delete
This PR is to switch merge and delete to no-chop merge and no-chop and
no-split deletions. There are some follow on tasks to do still but the
main functionality is complete.

The goal here is to prevent having to run chop compactions when merging
(or deleting) tablets which can be quite slow. The other goal is to get
rid of having to split tablets during deletions to simplify the manager
code. Instead, the changes here allow RFiles to be fenced off by by
valid ranges so that we don't need to compact/split anymore and when
scanning only the valid range is visible. Files can then be cleaned
up later with the normal compaction process.
This commit contains changes from the following PRs:

https://github.com/apache/accumulo/pull/3418
https://github.com/apache/accumulo/pull/3480
https://github.com/apache/accumulo/pull/3614
https://github.com/apache/accumulo/pull/3621
https://github.com/apache/accumulo/pull/3640
https://github.com/apache/accumulo/pull/3710
https://github.com/apache/accumulo/pull/3728