No-chop merge and delete (#3761)
This commit implements no-chop merge and no-chop and no-split deletions. 
There are some follow on tasks to do still but the main functionality is complete.

This change updates merge so chop compactions are no longer used when
when merging (or deleting) tablets, as that can be quite slow. The other change
is that splitting tablets during deletions is no longer necessary or used to 
simplify the manager code. Instead, the changes here allow RFiles to be fenced
off by by valid ranges so that we don't need to compact/split anymore and when
scanning only the valid range is visible. Files can then be cleaned up later with 
the normal compaction process.

This commit contains changes from the following PRs:

https://github.com/apache/accumulo/pull/3418
https://github.com/apache/accumulo/pull/3480
https://github.com/apache/accumulo/pull/3614
https://github.com/apache/accumulo/pull/3621
https://github.com/apache/accumulo/pull/3640
https://github.com/apache/accumulo/pull/3710
https://github.com/apache/accumulo/pull/3728

This closes #1327 

---------

Co-authored-by: Keith Turner <kturner@apache.org>
Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>