removes CHOP from CompactionKind SPI type.
This change removes CompactionKind.CHOP from the SPI w/o a deprecation
cycle.  This is allowed under the rules for SPI changes.  The rationale
for this is if CHOP is left then code will need to be added to Accumulo
to avoid its use and fail if it's ever seen.  Also documentation will
need to be added to the SPI explaining that Accumulo should never pass
CHOP to any plugins ever again that ideally plugins should error if they
see it.

Removing it seems like the least confusing option.