Revert #3737's EOFException changes
This commit partially reverts b1b2557f949e9212a1b1ca9b65f2d66c01a69edb
from #3737, to restore the previous infinite retry behavior when an
EOFException occurs in the transport. While this EOFException can be an
indicator of a fatal exception that should not be retried, it also
occurs during transient network failures, where a retry should occur. It
is not possible to easily detect which of these two cases is ocurring.
Since transient network failures are routine, and the fatal exception
that caused the issue that #3737 tried to address has a workaround via
configuration, this commit restores the previous behavior that assumes
the exception is caused by a transient issue.

This fixes #3762

This fixes the issue in #3762, where the ManagerRepairsDualAssignmentIT
intentionally triggered a transient network failure, and therefore
triggered a scan to fail, causing the IT as a whole to fail.