Modify ClientTabletCacheImpl.findExtentsToHost end key comparison (#3812)
The change in #3429 modified ClientTabletCache so that it would
preload Tablets for Scanners. It appears this change may
have caused the test in #3809 to fail. I modified the comparison
of the range to the tablet end key using the same logic that
ThriftScanner uses and it fixed the test.

Fixes #3809