Disable failsafe plugin in accumulo-start module
After the changes in #3817 junit is no longer a dependency for the start
module and this causes errors when running int tests as failsafe gives
an error because Junit is not a dependency.

Adding back junit as a dependency also fails as there are no unit tests
to run currently so this disables the failsafe plugin inside the start
module. It can be re-enabled in the future if tests are added.