Disable failsafe plugin in start and compactor modules (#3829)
After the changes in #3817 junit is no longer a dependency for the start/compactor
modules and this causes errors when running int tests as failsafe gives
an error because Junit is not a dependency.

Adding back junit as a dependency also fails as there are no unit tests
to run currently so this disables the failsafe plugin inside the start/compactor
module. It can be re-enabled in the future if tests are added.