Change getFirstKey/getLastKey to getFirstRow/getLastRow (#3830)
This commit changes FileSKVIterator getFirstKey and getLastKey methods
to return the first and last row instead. These methods are only used
for splits and in that case only the row is relevant. This was done
because fenced RFiles will only return the row portion of a key when
these methods are called (if the fencing is being enforced) so this
keeps the API consistent.