Use OperationId to unhost tablets, separate from location being absent (#3835)
This commit uses the operationId to unhost tablets. This required
changing the logic that set the operationId to not require the
location to be empty. Transactions that set the operationId will
have to wait for the location to be empty before continuing and
will likely want to continue to verify that the location is empty
as the transaction progresses.

Fixes #3406