Modified FATE to allow transactions to run to completion without interruption
FATE processes transactions in an interleaved peice-wise manner. In the case
of a multi-step transaction, FATE could potentially execute a step of another
transaction before executing the next step of the current transaction. This
change enables the developer to indicate via a new method on the Repo interface
that the transaction type should be run start to finish without being interrupted.