WIP modify merge and deleterows to use conditional mutations
Added a new FATE repo that acquires operations ids on all of the tablets
in the merge range.  This causes the tablet group watcher to unload
tablets.  This new repo waits for all tablets to not have a location.

Modified the existing merge and delete rows FATE repos to use ample and
conditional mutations.  Using ample simplified the code.

Modified the delete rows FATE repo to only delete or fence files, but
not merge tablets.  Had the code call the merge code to delete tablets.
This code is working (its deleting the expected data) but its producing
different splits than before.  The code is much much simpler now that
delete rows is not merging tablets, need to figure out why the splits
differ.

Removed all merge code from tablet group watcher and related classes
that was related to unassigning merging tablets.  Setting the operation
ids takes care of this now.

Modified the final FATE repo to delete operation ids.

MergeIT and DeleteRowsIT are passing with this change, however had to
comment out the checks of expected splits in DeleteRowsIT because of the
issue mentioned earlier. Still looking into this issue.