fixes tablet location checks in tabletserver (#3891)
During write ahead log recovery the tablet server was trying to write a
new file and require the tablets current location to be the tablet
server.  However at that time the tablet server was the tablets future
location and therefore the update failed.  This commit adust the
expected location to be the future location during recovery.