Acquires tablet servers and their groups atomically
Fixes two bugs that could cause the set of tservers and map of group
tservers to have an inconsistent set of tservers.

Modified LiveTserverSet so that the set of tablet servers and the tablet
servers resource groups can be acquired atomically.  The code was
acquiring this information at two different times with two different lock
acquisitions, which could have led to race condtions resulting
differences in set and the map.

Also modified TGW.TabletLists to filter out shutting down servers from
the grouped tservers in addition to filtering from the tserver set.